+++ pets/src/main/java/com/pets/domain/service/PetService.java
@@ -13,22 +13,38 @@ public class PetService {
     @Autowired
     private PetRepository petRepository;
 
+    @Autowired
+    private ClientService clientService;
+
     public List<PetDTO> getAllPets() {
         return petRepository.getAllPets();
     }
 
     public Optional<PetDTO> getPetById(int petId) {
-        return petRepository.getPetById(petId);
+        Optional<PetDTO> pet = petRepository.getPetById(petId);
+        if (pet.isEmpty()) {
+            throw new IllegalArgumentException("No se encontró la mascota con el ID especificado.");
+        }
+        return pet;
     }
 
     public PetDTO savePet(PetDTO pet) {
+        if (clientService.getClientById(pet.getOwnerId()).isEmpty()) {
+            throw new IllegalArgumentException("El ID del cliente no es válido.");
+        }
+        if (!"Hembra".equalsIgnoreCase(pet.getGender()) && !"Macho".equalsIgnoreCase(pet.getGender())) {
+            throw new IllegalArgumentException("El género debe ser 'Hembra' o 'Macho'.");
+        }
         return petRepository.savePet(pet);
     }
 
     public boolean deletePet(int petId) {
-        return getPetById(petId).map(pet -> {
+        Optional<PetDTO> pet = getPetById(petId);
+        if (pet.isPresent()) {
             petRepository.deletePet(petId);
             return true;
-        }).orElse(false);
+        } else {
+            throw new IllegalArgumentException("No se encontró la mascota con el ID especificado.");
+        }
     }
 }
