+++ pets/src/main/java/com/pets/domain/service/ServiceService.java
@@ -13,22 +13,35 @@ public class ServiceService {
     @Autowired
     private ServiceRepository serviceRepository;
 
+    @Autowired
+    private CategoryService categoryService;
+
     public List<ServiceDTO> getAllServices() {
         return serviceRepository.getAllServices();
     }
 
     public Optional<ServiceDTO> getServiceById(int serviceId) {
-        return serviceRepository.getServiceById(serviceId);
+        Optional<ServiceDTO> service = serviceRepository.getServiceById(serviceId);
+        if (service.isEmpty()) {
+            throw new IllegalArgumentException("No se encontró el servicio con el ID especificado.");
+        }
+        return service;
     }
 
     public ServiceDTO saveService(ServiceDTO service) {
+        if (categoryService.getCategoryById(service.getCategoryId()).isEmpty()) {
+            throw new IllegalArgumentException("El ID de la categoría no existe.");
+        }
         return serviceRepository.saveService(service);
     }
 
     public boolean deleteService(int serviceId) {
-        return getServiceById(serviceId).map(service -> {
+        Optional<ServiceDTO> service = getServiceById(serviceId);
+        if (service.isPresent()) {
             serviceRepository.deleteService(serviceId);
             return true;
-        }).orElse(false);
+        } else {
+            throw new IllegalArgumentException("No se encontró el servicio con el ID especificado.");
+        }
     }
 }
