+++ pets/src/main/java/com/pets/domain/service/UserService.java
@@ -18,17 +18,27 @@ public class UserService {
     }
 
     public Optional<UserDTO> getUserById(int userId) {
-        return userRepository.getUserById(userId);
+        Optional<UserDTO> user = userRepository.getUserById(userId);
+        if (user.isEmpty()) {
+            throw new IllegalArgumentException("No se encontró el usuario con el ID especificado.");
+        }
+        return user;
     }
 
     public UserDTO saveUser(UserDTO user) {
+        if (userRepository.findByLogin(user.getLogin()).isPresent()) {
+            throw new IllegalArgumentException("El login ya está en uso.");
+        }
         return userRepository.saveUser(user);
     }
 
     public boolean deleteUser(int userId) {
-        return getUserById(userId).map(user -> {
+        Optional<UserDTO> user = getUserById(userId);
+        if (user.isPresent()) {
             userRepository.deleteUser(userId);
             return true;
-        }).orElse(false);
+        } else {
+            throw new IllegalArgumentException("No se encontró el usuario con el ID especificado.");
+        }
     }
 }
